moved data fetch to _app.tsx
